requirements.txt
pandas>=1.5.0
shioaji>=1.0.0
python-dotenv>=1.0.0
mplfinance==0.12.9b7
